+++ Game Jam/Assets/Scripts/Ennemy/EnnemyPatrol.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 92b03ba8bcf483548bc239e1511b6318
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
